03-domain-configuration: Use a variable to refer to the domain
The examples use a hard-coded zone and domain. The zone is quite
flexible and folks can very easily use `example-com` as the examples do.
However, having to replace the domain in each command can be a little
tedious. To make things easier for the reader, this sets a variable with
the domain in the beginning of the section and uses it throughout the
code snippets. This way, it's easier for folks to use the command line
snippets directly with less mistakes.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>